Fix response packet size
The size of the response packet was calculated by checking the length of
the receive buffer. The buffer was created with a fixed size of 10240,
so that was always the size of the response payload.

Now we store the value of bytes read from ReadFromUDP and pass it on to
createTestResponse so that we can create a response packet of the same
size as the request packet.